+++ client/src/components/common/genric/ResponseCard.css
